+++ .idea/sqldialects.xml
@@ -4,6 +4,7 @@
     <file url="file://$PROJECT_DIR$/migrations/00000000000000_diesel_initial_setup/down.sql" dialect="PostgreSQL" />
     <file url="file://$PROJECT_DIR$/migrations/00000000000000_diesel_initial_setup/up.sql" dialect="PostgreSQL" />
     <file url="file://$PROJECT_DIR$/migrations/2024-10-12-110218_create_equipment/up.sql" dialect="PostgreSQL" />
+    <file url="file://$PROJECT_DIR$/migrations/2024-10-15-181243_create_auth_users_table/up.sql" dialect="PostgreSQL" />
     <file url="PROJECT" dialect="PostgreSQL" />
   </component>
 </project>
