+++ src/core/mod.rs
@@ -1,3 +1,3 @@
 pub mod errors;
 pub mod db_connection;
-mod crypto;
+pub mod crypto;
