+++ src/features/auth/auth_user.rs
@@ -1,87 +1,25 @@
-use async_trait::async_trait;
-use axum_login::{AuthUser as AxumAuthUser, AuthnBackend, UserId};
-use serde::{Deserialize, Serialize};
-use tokio::task;
+use diesel::prelude::*;
+use axum_login::{
+    AuthUser,
+    secrecy::SecretVec,
+};
 
-#[derive(Clone, Serialize, Deserialize, FromRow)]
-pub struct AuthUser {
-    id: i64,
-    pub username: String,
-    password: String,
-}
-
-// Here we've implemented `Debug` manually to avoid accidentally logging the
-// password hash.
-impl std::fmt::Debug for AuthUser {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("User")
-            .field("id", &self.id)
-            .field("username", &self.username)
-            .field("password", &"[redacted]")
-            .finish()
-    }
-}
-
-impl AxumAuthUser for AuthUser {
-    type Id = i64;
-
-    fn id(&self) -> Self::Id {
+impl AuthUser<i32> for AuthUserModel {
+    fn get_id(&self) -> i32 {
         self.id
     }
 
-    fn session_auth_hash(&self) -> &[u8] {
-        self.password.as_bytes() // We use the password hash as the auth
-        // hash--what this means
-        // is when the user changes their password the
-        // auth session becomes invalid.
+    fn get_password_hash(&self) -> SecretVec<u8> {
+        SecretVec::new(self.password_hash.clone().into())
     }
 }
 
-// This allows us to extract the authentication fields from forms. We use this
-// to authenticate requests with the backend.
-#[derive(Debug, Clone, Deserialize)]
-pub struct Credentials {
-    pub username: String,
-    pub password: String,
-    pub next: Option<String>,
+#[derive(Debug, Clone, Queryable, Selectable, Identifiable)]
+#[diesel(table_name = crate::schema::auth_users)]
+#[diesel(check_for_backend(diesel::pg::Pg))]
+pub struct AuthUserModel {
+    pub id: i32,
+    pub user_id: i32,
+    pub name: String,
+    pub password_hash: String,
 }
-
-#[async_trait]
-impl AuthnBackend for Backend {
-    type User = AuthUser;
-    type Credentials = Credentials;
-    type Error = Error;
-
-    async fn authenticate(
-        &self,
-        creds: Self::Credentials,
-    ) -> Result<Option<Self::User>, Self::Error> {
-        let user: Option<Self::User> = sqlx::query_as("select * from users where username = ? ")
-            .bind(creds.username)
-            .fetch_optional(&self.db)
-            .await?;
-
-        // Verifying the password is blocking and potentially slow, so we'll do so via
-        // `spawn_blocking`.
-        task::spawn_blocking(|| {
-            // We're using password-based authentication--this works by comparing our form
-            // input with an argon2 password hash.
-            Ok(user.filter(|user| verify_password(creds.password, &user.password).is_ok()))
-        })
-            .await?
-    }
-
-    async fn get_user(&self, user_id: &UserId<Self>) -> Result<Option<Self::User>, Self::Error> {
-        let user = sqlx::query_as("select * from users where id = ?")
-            .bind(user_id)
-            .fetch_optional(&self.db)
-            .await?;
-
-        Ok(user)
-    }
-}
-
-// We use a type alias for convenience.
-//
-// Note that we've supplied our concrete backend here.
-pub type AuthSession = axum_login::AuthSession<Backend>;
