+++ src/features/auth/mod.rs
@@ -1,2 +1,3 @@
 pub mod auth;
-pub mod auth_user;
+pub mod auth_user_bak;
+mod auth_user;
