+++ src/schema.rs
@@ -1,6 +1,17 @@
 // @generated automatically by Diesel CLI.
 
 diesel::table! {
+    auth_users (id) {
+        id -> Int4,
+        user_id -> Int4,
+        username -> Varchar,
+        password_hash -> Varchar,
+        created_at -> Timestamp,
+        updated_at -> Timestamp,
+    }
+}
+
+diesel::table! {
     equipments (id) {
         id -> Int4,
         publicid -> Uuid,
@@ -20,7 +31,10 @@ diesel::table! {
     }
 }
 
+diesel::joinable!(auth_users -> users (user_id));
+
 diesel::allow_tables_to_appear_in_same_query!(
+    auth_users,
     equipments,
     users,
 );
